+++ coursemanagement/src/main/java/com/example/coursemanagement/service/ProgramService.java
@@ -1,16 +1,16 @@
 package com.example.coursemanagement.service;
 
+import java.util.List;
+import java.util.stream.Collectors;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
 import com.example.coursemanagement.dto.CourseDTO;
 import com.example.coursemanagement.dto.ProgramDto;
 import com.example.coursemanagement.model.Course;
 import com.example.coursemanagement.model.Program;
 import com.example.coursemanagement.repository.ProgramRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
 
 @Service
 public class ProgramService {
@@ -23,21 +23,7 @@ public class ProgramService {
     }
 
     // Method to fetch program details by programId
-//    public Program getProgramById(Long programId) {
-//        // Check if program exists by programId
-//        Optional<Program> program = programRepository.findById(programId);
-//        if (program.isPresent()) {
-//            return program.get(); // Return the Program details if found
-//        } else {
-//            throw new RuntimeException("Program not found with ID: " + programId); // Handle case where Program is not found
-//        }
-//    }
-//
-//    public List<Program> getAllPrograms() {
-//        return programRepository.findAll(); // Fetches all programs from the database
-//    }
-
-    public ProgramDto getProgramById(Long programId) {
+    public ProgramDto getProgramById(Integer programId) {
         return programRepository.findById(programId)
                 .map(this::mapToDto)
                 .orElseThrow(() -> new RuntimeException("Program not found with ID: " + programId));
